Refactor: extract actor engine to actor.rs, clean up sim.rs
- Extract actor-model simulation (NodeListWrapper, ActorSimulation,
  init_nodes, run logic) into sim/actor.rs
- sim.rs is now a thin dispatch layer: Simulation newtype wrapping
  SimulationInner enum (Actor vs Sequential)
- Unify sequential single-shard and multi-shard builders into a
  single build_typed() function
- Group cross-shard state into CrossShardState sub-struct
- build() takes event_sender directly, creates its own infrastructure
- Remove dead code (per_shard_node_configs, init_node_impls)
- Net -407 lines across sim.rs + sequential.rs
