+++ sim-rs/sim-core/src/sim.rs
@@ -1,27 +1,18 @@
-use std::{collections::HashMap, sync::Arc, time::Duration};
+use std::{sync::Arc, time::Duration};
 
 use anyhow::Result;
-use futures::future::BoxFuture;
-use rand::RngCore;
-use rand_chacha::{ChaChaRng, rand_core::SeedableRng};
-use slot::SlotWitness;
-use tokio::{select, sync::mpsc, task::JoinSet};
+use rand_chacha::ChaChaRng;
+use tokio::sync::mpsc;
 use tokio_util::sync::CancellationToken;
 
 use crate::{
-    clock::{Clock, ClockCoordinator, Timestamp},
-    config::{CpuTimeConfig, LeiosVariant, NodeConfiguration, NodeId, SimConfiguration},
+    clock::{Clock, Timestamp},
+    config::{CpuTimeConfig, NodeConfiguration, NodeId, SimConfiguration},
     events::EventTracker,
     model::Transaction,
-    network::{Network, ShardLookup},
-    sharding::shard::{Shard, build_shards},
 };
 
-use self::{
-    driver::NodeDriver, leios::LeiosNode, linear_leios::LinearLeiosNode,
-    stracciatella::StracciatellaLeiosNode,
-};
-
+pub(crate) mod actor;
 mod cpu;
 mod driver;
 mod leios;
@@ -34,55 +25,14 @@ mod stracciatella;
 mod tests;
 pub(crate) mod tx;
 
-enum NodeListWrapper {
-    Leios(Vec<NodeDriver<LeiosNode>>),
-    Stracciatella(Vec<NodeDriver<StracciatellaLeiosNode>>),
-    LinearLeios(Vec<NodeDriver<LinearLeiosNode>>),
-}
-impl NodeListWrapper {
-    fn run_all(&mut self, set: &mut JoinSet<Result<()>>) {
-        match self {
-            Self::Leios(nodes) => {
-                for node in nodes.drain(..) {
-                    set.spawn(node.run());
-                }
-            }
-            Self::Stracciatella(nodes) => {
-                for node in nodes.drain(..) {
-                    set.spawn(node.run());
-                }
-            }
-            Self::LinearLeios(nodes) => {
-                for node in nodes.drain(..) {
-                    set.spawn(node.run());
-                }
-            }
-        }
-    }
-}
+// Re-export for the attacker module
+pub(crate) use actor::{Actor, ActorInitArgs};
 
-trait Actor {
-    fn run(self: Box<Self>) -> BoxFuture<'static, Result<()>>;
-}
+pub struct Simulation(SimulationInner);
 
-struct ActorInitArgs<'a, N: NodeImpl> {
-    pub config: Arc<SimConfiguration>,
-    pub clock: Clock,
-    pub tracker: EventTracker,
-    pub nodes: &'a mut [N],
-}
-
-fn no_additional_actors<N: NodeImpl>(_args: ActorInitArgs<N>) -> Vec<Box<dyn Actor>> {
-    vec![]
-}
-
-pub struct Simulation {
-    shards: Vec<Shard>,
-    slot_witness: SlotWitness,
-    nodes: NodeListWrapper,
-    actors: Vec<Box<dyn Actor>>,
-    /// If set, the sequential DES engine handles execution instead of the actor model.
-    sequential: Option<Box<dyn sequential::SequentialRunner>>,
+enum SimulationInner {
+    Actor(actor::ActorSimulation),
+    Sequential(Box<dyn sequential::SequentialRunner>),
 }
 
 impl Simulation {
@@ -90,246 +40,30 @@ impl Simulation {
         config: SimConfiguration,
         event_sender: mpsc::UnboundedSender<(crate::events::Event, Timestamp)>,
     ) -> Result<Self> {
-        let use_sequential = config.sequential_engine;
         let config = Arc::new(config);
-        let shard_count = config.shard_count;
 
-        // Build shard lookup: node_id -> shard_index
-        let shard_lookup: ShardLookup = crate::sharding::compute_shard_lookup(&config);
-
-        // Create per-shard clock coordinators
-        let mut clock_coordinators: Vec<ClockCoordinator> = (0..shard_count)
-            .map(|_| ClockCoordinator::new(config.timestamp_resolution))
-            .collect();
-
-        let clocks: Vec<Clock> = clock_coordinators.iter().map(|cc| cc.clock()).collect();
-
-        // Create per-shard EventTrackers (all share the same sender channel)
-        let trackers: Vec<EventTracker> = clocks
-            .iter()
-            .map(|clock| EventTracker::new(event_sender.clone(), clock.clone(), &config.nodes))
-            .collect();
-
-        if use_sequential {
+        if config.sequential_engine {
             tracing::info!("using sequential DES engine");
-            let seq = sequential::build(
-                config.clone(),
-                trackers[0].clone(),
-                clocks[0].clone(),
-                clock_coordinators[0].shared_time(),
-            );
-            let slot_witness =
-                SlotWitness::new(clocks[0].barrier(), trackers[0].clone(), &config);
-            return Ok(Self {
-                shards: vec![],
-                slot_witness,
-                nodes: NodeListWrapper::Leios(vec![]),
-                actors: vec![],
-                sequential: Some(seq),
-            });
+            return Ok(Self(SimulationInner::Sequential(sequential::build(
+                config,
+                event_sender,
+            ))));
         }
 
-        let (nodes, actors, shards) = match config.variant {
-            LeiosVariant::Linear | LeiosVariant::LinearWithTxReferences => {
-                let (nodes, actors, tx_sinks, networks) = Self::init_nodes(
-                    &config, &trackers, &clocks, &shard_lookup,
-                    NodeListWrapper::LinearLeios,
-                    linear_leios::register_actors,
-                )?;
-                let shards = build_shards(
-                    &config, &clocks, &mut clock_coordinators, &shard_lookup,
-                    tx_sinks, networks,
-                );
-                (nodes, actors, shards)
-            }
-            LeiosVariant::FullWithoutIbs => {
-                let (nodes, actors, tx_sinks, networks) = Self::init_nodes(
-                    &config, &trackers, &clocks, &shard_lookup,
-                    NodeListWrapper::Stracciatella,
-                    no_additional_actors,
-                )?;
-                let shards = build_shards(
-                    &config, &clocks, &mut clock_coordinators, &shard_lookup,
-                    tx_sinks, networks,
-                );
-                (nodes, actors, shards)
-            }
-            _ => {
-                let (nodes, actors, tx_sinks, networks) = Self::init_nodes(
-                    &config, &trackers, &clocks, &shard_lookup,
-                    NodeListWrapper::Leios,
-                    no_additional_actors,
-                )?;
-                let shards = build_shards(
-                    &config, &clocks, &mut clock_coordinators, &shard_lookup,
-                    tx_sinks, networks,
-                );
-                (nodes, actors, shards)
-            }
-        };
-
-        // SlotWitness uses shard 0's clock
-        let slot_witness = SlotWitness::new(clocks[0].barrier(), trackers[0].clone(), &config);
-
-        Ok(Self {
-            shards,
-            slot_witness,
-            nodes,
-            actors,
-            sequential: None,
-        })
+        Ok(Self(SimulationInner::Actor(actor::ActorSimulation::new(
+            config,
+            event_sender,
+        )?)))
     }
 
-    #[allow(clippy::type_complexity, clippy::too_many_arguments)]
-    fn init_nodes<N, NLF, AAF>(
-        config: &Arc<SimConfiguration>,
-        trackers: &[EventTracker],
-        clocks: &[Clock],
-        shard_lookup: &ShardLookup,
-        node_list_wrapper_fn: NLF,
-        additional_actors_fn: AAF,
-    ) -> Result<(
-        NodeListWrapper,
-        Vec<Box<dyn Actor>>,
-        Vec<HashMap<NodeId, mpsc::UnboundedSender<Arc<Transaction>>>>,
-        Vec<Network<MiniProtocol, N::Message>>,
-    )>
-    where
-        N: NodeImpl,
-        N::Message: Send + 'static,
-        NLF: FnOnce(Vec<NodeDriver<N>>) -> NodeListWrapper,
-        AAF: FnOnce(ActorInitArgs<N>) -> Vec<Box<dyn Actor>>,
-    {
-        let shard_count = clocks.len();
-        let mut rng = ChaChaRng::seed_from_u64(config.seed);
-
-        // Create per-shard networks
-        let mut networks: Vec<Network<MiniProtocol, N::Message>> = clocks
-            .iter()
-            .map(|clock| Network::new(clock.clone()))
-            .collect();
-
-        // Create node implementations
-        let mut node_impls = vec![];
-        for node_config in &config.nodes {
-            let shard = shard_lookup[&node_config.id];
-            node_impls.push(N::new(
-                node_config,
-                config.clone(),
-                trackers[shard].clone(),
-                ChaChaRng::seed_from_u64(rng.next_u64()),
-                clocks[shard].clone(),
-            ));
-        }
-
-        // Additional actors (e.g. linear_leios attackers) use shard 0's clock
-        let actor_args = ActorInitArgs {
-            config: config.clone(),
-            clock: clocks[0].clone(),
-            tracker: trackers[0].clone(),
-            nodes: &mut node_impls,
-        };
-        let actors = additional_actors_fn(actor_args);
-
-        // Create node drivers, grouped by shard
-        let mut per_shard_tx_sinks: Vec<HashMap<NodeId, mpsc::UnboundedSender<Arc<Transaction>>>> =
-            (0..shard_count).map(|_| HashMap::new()).collect();
-        let mut all_nodes = vec![];
-        for (node_config, node) in config.nodes.iter().zip(node_impls) {
-            let id = node_config.id;
-            let shard = shard_lookup[&id];
-            let (tx_sink, tx_source) = mpsc::unbounded_channel();
-            per_shard_tx_sinks[shard].insert(id, tx_sink);
-            let driver = NodeDriver::new(
-                node,
-                node_config,
-                config.clone(),
-                &mut networks[shard],
-                tx_source,
-                trackers[shard].clone(),
-                clocks[shard].barrier(),
-            );
-            all_nodes.push(driver);
-        }
-
-        Ok((
-            node_list_wrapper_fn(all_nodes),
-            actors,
-            per_shard_tx_sinks,
-            networks,
-        ))
-    }
-
-    /// Run the simulation. Consumes self so shards can be spawned as independent tasks.
-    pub async fn run(mut self, token: CancellationToken) -> Result<()> {
-        // Sequential engine path
-        if let Some(seq) = self.sequential.take() {
-            let result = tokio::task::spawn_blocking(move || seq.run(token)).await?;
-            return result;
-        }
-
-        let mut set = JoinSet::new();
-
-        self.nodes.run_all(&mut set);
-        for actor in self.actors {
-            set.spawn(actor.run());
-        }
-
-        if self.shards.len() == 1 {
-            // Single shard: run in the current task via select! for minimal overhead
-            let shard = self.shards.pop().unwrap();
-            select! {
-                biased;
-                _ = token.cancelled() => {}
-                _ = self.slot_witness.run() => {}
-                result = shard.run(token.clone()) => { result? }
-                result = set.join_next() => { result.unwrap()??; }
-            };
-        } else {
-            // Multi-shard: spawn each shard as its own tokio task (independent LP)
-            let mut slot_witness = self.slot_witness;
-            let slot_token = token.clone();
-            set.spawn(async move {
-                select! {
-                    _ = slot_token.cancelled() => {}
-                    _ = slot_witness.run() => { slot_token.cancel(); }
-                }
-                Ok(())
-            });
-
-            for shard in self.shards {
-                let shard_token = token.clone();
-                set.spawn(shard.run(shard_token));
+    pub async fn run(self, token: CancellationToken) -> Result<()> {
+        match self.0 {
+            SimulationInner::Sequential(seq) => {
+                tokio::task::spawn_blocking(move || seq.run(token)).await?
             }
-
-            // Wait for tasks — errors during normal operation propagate,
-            // but errors after cancellation (e.g. closed channels) are expected.
-            while let Some(result) = set.join_next().await {
-                match result {
-                    Ok(Ok(())) => {}
-                    Ok(Err(e)) => {
-                        if !token.is_cancelled() {
-                            token.cancel();
-                            return Err(e);
-                        }
-                    }
-                    Err(e) => {
-                        if !token.is_cancelled() {
-                            token.cancel();
-                            return Err(e.into());
-                        }
-                    }
-                }
-            }
+            SimulationInner::Actor(actor) => actor.run(token).await,
         }
-
-        Ok(())
     }
-}
-
-trait SimMessage: Clone + std::fmt::Debug {
-    fn protocol(&self) -> MiniProtocol;
-    fn bytes_size(&self) -> u64;
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, Hash)]
@@ -341,13 +75,18 @@ pub enum MiniProtocol {
     Vote,
 }
 
-trait SimCpuTask {
+pub(crate) trait SimMessage: Clone + std::fmt::Debug {
+    fn protocol(&self) -> MiniProtocol;
+    fn bytes_size(&self) -> u64;
+}
+
+pub(crate) trait SimCpuTask {
     fn name(&self) -> String;
     fn extra(&self) -> String;
     fn times(&self, config: &CpuTimeConfig) -> Vec<Duration>;
 }
 
-struct EventResult<N: NodeImpl> {
+pub(crate) struct EventResult<N: NodeImpl> {
     messages: Vec<(NodeId, N::Message)>,
     tasks: Vec<N::Task>,
     timed_events: Vec<(Timestamp, N::TimedEvent)>,
@@ -384,7 +123,7 @@ impl<N: NodeImpl> EventResult<N> {
     }
 }
 
-trait NodeImpl: Sized {
+pub(crate) trait NodeImpl: Sized {
     type Message: SimMessage;
     type Task: SimCpuTask;
     type TimedEvent;
